session/window/pane options: fix restoring empty values and no values
When option value is empty, tmux show-option prints it as option-name
and '', but when it has a value, it's printed in double-quotes instead.
Rework seds to strip single quotes too.

Second thing, when there are no options in save file, the while loop
gets executed once with empty variables, but tmux does not like to set
option '' to ''.